// src/app/shared/interface/agency-email-preview-data.interface.ts
export interface AgencyEmailPreviewDataInterface {
  template: string;
  data: {
    propertyData: string[];
    background_color_head_foot: string;
    background_color_main: string;
    background_color_message: string;
    background_color_property_title: string;
    text_color_head_foot: string;
    text_color_main: string;
    text_color_message: string;
    text_color_property_title: string;
  };
}